# backend/gateway_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway банковского приложения. Единая точка входа для клиентских запросов.",
        "title": "Gateway Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/customers/me": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает полную агрегированную информацию о текущем пользователе (ФИО, Паспорт, Контакты и т.д.).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Получить профиль клиента",
                "responses": {
                    "200": {
                        "description": "Полный профиль пользователя",
                        "schema": {
                            "$ref": "#/definitions/schemas.CustomerProfileDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Помечает профиль как удалённый (Soft Delete) и завершает все активные сессии.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Удалить профиль клиента",
                "responses": {
                    "200": {
                        "description": "Профиль успешно удален",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/customers/me/contacts": {
            "patch": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Позволяет обновить Email или номер телефона.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Смена контактных данных",
                "parameters": [
                    {
                        "description": "Новые контакты",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.ContactsUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Контакты успешно обновлены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/customers/me/passport": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Обновляет паспортные данные пользователя (например, при замене документа).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Смена паспорта",
                "parameters": [
                    {
                        "description": "Новые паспортные данные",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PassportPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Паспорт успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/customers/me/personal-data": {
            "patch": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Позволяет изменить имя, фамилию или отчество.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Обновить ФИО",
                "parameters": [
                    {
                        "description": "Новые анкетные данные",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PersonalDataUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные успешно обновлены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/onboarding": {
            "post": {
                "description": "Инициирует процесс создания нового клиента.\nСоздаёт временный профиль и выдаёт X-Onboarding-Token для прохождения последующих шагов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Начать регистрацию (Шаг 0)",
                "responses": {
                    "201": {
                        "description": "Регистрация начата",
                        "schema": {
                            "$ref": "#/definitions/schemas.OnboardingStartResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/completion": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Переносит данные из временного хранилища в основной профиль и активирует аккаунт.\nВ ответе возвращается сессионный токен для немедленного входа.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Завершение регистрации",
                "responses": {
                    "200": {
                        "description": "Регистрация успешно завершена",
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Не все шаги пройдены",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/contacts": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Сохраняет контактные данные для связи и уведомлений.\nТребует заголовок X-Onboarding-Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Шаг 4: Контакты (Email/Телефон)",
                "parameters": [
                    {
                        "description": "Email и Телефон",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.ContactsPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Контактные данные сохранены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/email/send": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Отправляет проверочный код на Email, указанный на предыдущем шаге.\nТребует заголовок X-Onboarding-Token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Запросить Email-подтверждение",
                "responses": {
                    "200": {
                        "description": "Код успешно отправлен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/email/verify": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Проверяет 6-значный код, отправленный на почту.\nТребует заголовок X-Onboarding-Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Подтвердить Email",
                "parameters": [
                    {
                        "description": "Код подтверждения",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.VerifyEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email подтвержден",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный или просроченный код",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/identifiers": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Сохраняет государственные идентификаторы.\nТребует заголовок X-Onboarding-Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Шаг 3: ИНН и СНИЛС",
                "parameters": [
                    {
                        "description": "ИНН и СНИЛС",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.IdentifiersPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные сохранены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации (дубликат или формат)",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/passport": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Сохраняет данные паспорта РФ.\nТребует заголовок X-Onboarding-Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Шаг 2: Паспортные данные",
                "parameters": [
                    {
                        "description": "Данные паспорта",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PassportPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные сохранены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный токен",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/onboarding/personal-data": {
            "post": {
                "security": [
                    {
                        "OnboardingToken": []
                    }
                ],
                "description": "Сохраняет базовые анкетные данные во временный профиль.\nТребует заголовок X-Onboarding-Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Шаг 1: ФИО и дата рождения",
                "parameters": [
                    {
                        "description": "Анкетные данные",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PersonalDataPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные сохранены",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации данных",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный токен онбординга",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ContactsUpdate": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Новый Email",
                    "type": "string",
                    "format": "email",
                    "example": "new@example.com"
                },
                "phone": {
                    "description": "Новый номер телефона",
                    "type": "string",
                    "format": "phone",
                    "example": "+79990001122"
                }
            }
        },
...
        "schemas.CustomerProfileDTO": {
            "type": "object",
            "properties": {
                "contacts": {
                    "description": "Контактные данные",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.ContactsPayload"
                        }
                    ]
                },
                "id": {
                    "description": "Уникальный идентификатор пользователя (UUID)",
                    "type": "string",
                    "format": "uuid",
                    "example": "user-uuid"
                },
                "identifiers": {
                    "description": "ИНН и СНИЛС",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.IdentifiersPayload"
                        }
                    ]
                },
                "passport": {
                    "description": "Паспортные данные",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.PassportPayload"
                        }
                    ]
                },
                "personal_data": {
                    "description": "Анкетные данные (ФИО, ДР, Пол)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.PersonalDataPayload"
                        }
                    ]
                },
                "status": {
                    "description": "Статус учетной записи",
                    "type": "string",
                    "enum": [
                        "active",
                        "blocked",
                        "deleted"
                    ],
                    "example": "active"
                }
            }
        },
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Подробное сообщение об ошибке для разработчика и пользователя",
                    "type": "string",
                    "example": "Неверный PIN-код или формат данных"
                }
            }
        },
...
        "schemas.PersonalDataUpdate": {
            "type": "object",
            "properties": {
                "first_name": {
                    "description": "Новое имя",
                    "type": "string",
                    "example": "Петр"
                },
                "last_name": {
                    "description": "Новая фамилия",
                    "type": "string",
                    "example": "Петров"
                },
                "middle_name": {
                    "description": "Новое отчество",
                    "type": "string",
                    "example": "Петрович"
                }
            }
        },
...
        "schemas.SuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Статус успеха (обычно 'success')",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OnboardingToken": {
            "description": "Токен регистрации, полученный на Шаге 0 (Start Onboarding)",
            "type": "apiKey",
            "name": "X-Onboarding-Token",
            "in": "header"
        },
        "SessionToken": {
            "description": "Сессионный токен, полученный при авторизации",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    },
...
}